New email alert options for monitor
Now an email alert is only send the first time a request fails or is out
of order.
An email is now send also send when request works again.

An extra option is added for getting alert email when a request is out
of order (The request took to long time to fulfill)